+++ src/containers/ContactDetailsPage/ContactDetailsPage.duck.js
@@ -12,6 +12,10 @@ export const SAVE_PHONE_NUMBER_ERROR = 'app/ContactDetailsPage/SAVE_PHONE_NUMBER
 
 export const SAVE_CONTACT_DETAILS_CLEAR = 'app/ContactDetailsPage/SAVE_CONTACT_DETAILS_CLEAR';
 
+export const RESET_PASSWORD_REQUEST = 'app/ContactDetailsPage/RESET_PASSWORD_REQUEST';
+export const RESET_PASSWORD_SUCCESS = 'app/ContactDetailsPage/RESET_PASSWORD_SUCCESS';
+export const RESET_PASSWORD_ERROR = 'app/ContactDetailsPage/RESET_PASSWORD_ERROR';
+
 // ================ Reducer ================ //
 
 const initialState = {
@@ -19,6 +23,8 @@ const initialState = {
   savePhoneNumberError: null,
   saveContactDetailsInProgress: false,
   contactDetailsChanged: false,
+  resetPasswordInProgress: false,
+  resetPasswordError: null,
 };
 
 export default function reducer(state = initialState, action = {}) {
@@ -48,6 +54,14 @@ export default function reducer(state = initialState, action = {}) {
         contactDetailsChanged: false,
       };
 
+    case RESET_PASSWORD_REQUEST:
+      return { ...state, resetPasswordInProgress: true, resetPasswordError: null };
+    case RESET_PASSWORD_SUCCESS:
+      return { ...state, resetPasswordInProgress: false };
+    case RESET_PASSWORD_ERROR:
+      console.error(payload); // eslint-disable-line no-console
+      return { ...state, resetPasswordInProgress: false, resetPasswordError: payload };
+
     default:
       return state;
   }
@@ -69,6 +83,16 @@ export const savePhoneNumberError = error => ({
 });
 
 export const saveContactDetailsClear = () => ({ type: SAVE_CONTACT_DETAILS_CLEAR });
+
+export const resetPasswordRequest = () => ({ type: RESET_PASSWORD_REQUEST });
+
+export const resetPasswordSuccess = () => ({ type: RESET_PASSWORD_SUCCESS });
+
+export const resetPasswordError = e => ({
+  type: RESET_PASSWORD_ERROR,
+  error: true,
+  payload: e,
+});
 
 // ================ Thunks ================ //
 
@@ -216,4 +240,12 @@ export const saveContactDetails = params => (dispatch, getState, sdk) => {
   } else if (phoneNumberChanged) {
     return dispatch(savePhoneNumber({ phoneNumber }));
   }
+};
+
+export const resetPassword = email => (dispatch, getState, sdk) => {
+  dispatch(resetPasswordRequest());
+  return sdk.passwordReset
+    .request({ email })
+    .then(() => dispatch(resetPasswordSuccess()))
+    .catch(e => dispatch(resetPasswordError(storableError(e))));
 };